Move current user to root folder
